pyproject.toml
[project]
name = "databricks-npx-mcp-remote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b251007",
    "azure-identity>=1.25.1",
    "python-dotenv>=1.1.1",
]
